SYNC-1540 Device registration

+++ modules/util/sync-engine/src/main/java/com/liferay/sync/engine/service/SyncFileService.java
@@ -248,6 +248,10 @@ public class SyncFileService {
 							syncFile.getFilePathName());
 
 					for (SyncFile childSyncFile : childSyncFiles) {
+						if (childSyncFile.isSystem()) {
+							continue;
+						}
+
 						childSyncFile.setUiEvent(syncFile.getUiEvent());
 
 						doDeleteSyncFile(childSyncFile, notify);
